storybook: move care ui elements to care ui dir
